Added flag to enable/disable cron jobs

+++ manifests/agent/schedule.pp
@@ -5,6 +5,7 @@
 # @example
 #   include puppet::agent::schedule
 class puppet::agent::schedule (
+  Boolean $enable               = true,
   String  $job_name             = 'puppet agent run',
   Array[String]
           $job_arguments        = [
@@ -23,18 +24,20 @@ class puppet::agent::schedule (
   $agent_run_minute = fqdn_rand(60, $job_name)
   $agent_run_arguments = join($job_arguments, ' ')
 
-  cron { $job_name:
-    command     => "/opt/puppetlabs/bin/puppet agent ${agent_run_arguments}",
-    hour        => absent,
-    minute      => $agent_run_minute,
-    month       => absent,
-    monthday    => absent,
-    user        => 'root',
-    weekday     => absent,
-    environment => 'PATH=/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin',
+  if $enable {
+    cron { $job_name:
+      command     => "/opt/puppetlabs/bin/puppet agent ${agent_run_arguments}",
+      hour        => absent,
+      minute      => $agent_run_minute,
+      month       => absent,
+      monthday    => absent,
+      user        => 'root',
+      weekday     => absent,
+      environment => 'PATH=/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin',
+    }
   }
 
-  if $reboot_job {
+  if $enable and $reboot_job {
     cron { "${job_name} on boot":
       command     => "/opt/puppetlabs/bin/puppet agent ${agent_run_arguments}",
       special     => 'reboot',
